+++ src/KalmanFilter.h
@@ -11,7 +11,7 @@ class KalmanFilter
 {
 public:
   void Yiorderfilter(float angle_m, float gyro_m, float dt, float K1);
-  void Kalman_Filter(double angle_m, double gyro_m, float dt, float Q_angle, float Q_gyro, float R_angle, float C_0);
+  void Kalman_Filter(float angle_m, float gyro_m, float dt, float Q_angle, float Q_gyro, float R_angle, float C_0);
   void Angle(int16_t ax, int16_t ay, int16_t az, int16_t gx, int16_t gy, int16_t gz, float dt, float Q_angle, float Q_gyro, float R_angle, float C_0, float K1);
   float Gyro_x, Gyro_y, Gyro_z;
   float accelz = 0;
